+++ docs/examples/01_pet_geometry/02_run_regular_polygon_pet_sino.py
@@ -7,11 +7,6 @@ between two LOR endpoints can be sorted into a sinogram containing a radial,
 view and plane dimension.
 This example shows how this can be done using the :class:`.RegularPolygonPETLORDescriptor`
 
-.. note::
-
-   To run this example locally, download
-   `example_utils.py <https://raw.githubusercontent.com/KUL-recon-lab/parallelproj/main/docs/examples/example_utils.py>`_
-   into the **same folder** as this script. Make sure ``parallelproj`` is installed.
 """
 
 # %%
@@ -21,7 +16,8 @@ import parallelproj.pet_lors
 import matplotlib.pyplot as plt
 
 # %%
-from example_utils import suggest_array_backend_and_device
+
+from parallelproj._examples_utils import suggest_array_backend_and_device
 
 # To use a specific backend and/or device, replace the None arguments, e.g.:
 #   xp, dev = suggest_array_backend_and_device(backend="numpy", dev="cpu") or by setting xp and dev manually