+++ docs/examples/01_pet_geometry/05_run_zigzag_comparison.py
@@ -20,11 +20,6 @@ parameter of :class:`.RegularPolygonPETLORDescriptor`.
 This example visualises both conventions for a minimal scanner with 1 ring and
 8 detector endpoints.
 
-.. note::
-
-   To run this example locally, download
-   `example_utils.py <https://raw.githubusercontent.com/KUL-recon-lab/parallelproj/main/docs/examples/example_utils.py>`_
-   into the **same folder** as this script. Make sure ``parallelproj`` is installed.
 """
 
 # %%
@@ -34,7 +29,8 @@ import parallelproj.pet_scanners
 import parallelproj.pet_lors
 
 # %%
-from example_utils import suggest_array_backend_and_device
+
+from parallelproj._examples_utils import suggest_array_backend_and_device
 
 xp, dev = suggest_array_backend_and_device(None, None)