+++ docs/examples/02_pet_sinogram_projections/04_run_pet_listmode_projector.py
@@ -7,16 +7,12 @@ for both non-TOF and TOF acquisitions, including geometrical forward
 projection in listmode, image-based resolution model, and a listmode
 attenuation model.
 
-.. note::
-
-   To run this example locally, download
-   `example_utils.py <https://raw.githubusercontent.com/KUL-recon-lab/parallelproj/main/docs/examples/example_utils.py>`_
-   into the **same folder** as this script. Make sure ``parallelproj`` is installed.
 """
 
 # %%
 import matplotlib.pyplot as plt
-from example_utils import show_vol_cuts
+
+from parallelproj._examples_utils import show_vol_cuts
 
 import parallelproj.pet_scanners
 import parallelproj.projectors
@@ -24,7 +20,7 @@ import parallelproj.operators
 import parallelproj.tof
 
 # %%
-from example_utils import suggest_array_backend_and_device
+from parallelproj._examples_utils import suggest_array_backend_and_device
 
 # To use a specific backend and/or device, replace the None arguments, e.g.:
 #   xp, dev = suggest_array_backend_and_device(backend="numpy", dev="cpu") or by setting xp and dev manually