+++ docs/examples/03_algorithms/00_run_mlem_osem_svrg.py
@@ -25,11 +25,6 @@ subject to :math:`x \\geq 0`:
     convergence plot therefore understates the true computational cost of
     SVRG relative to OSEM by a factor of roughly two.
 
-.. note::
-
-   To run this example locally, download
-   `example_utils.py <https://raw.githubusercontent.com/KUL-recon-lab/parallelproj/main/docs/examples/example_utils.py>`_
-   into the **same folder** as this script. Make sure ``parallelproj`` is installed.
 """
 
 # %%
@@ -47,11 +42,12 @@ import parallelproj.projectors
 from parallelproj import to_numpy_array, Array
 from parallelproj.functions import NegPoissonLogL, C2AffineObjective, C1Function
 
-from example_utils import show_vol_cuts
-from example_utils import elliptic_cylinder_phantom
+
+from parallelproj._examples_utils import show_vol_cuts
+from parallelproj._examples_utils import elliptic_cylinder_phantom
 
 # %%
-from example_utils import suggest_array_backend_and_device
+from parallelproj._examples_utils import suggest_array_backend_and_device
 
 # To use a specific backend and/or device, replace the None arguments, e.g.:
 #   xp, dev = suggest_array_backend_and_device(backend="numpy", dev="cpu") or by setting xp and dev manually