+++ docs/examples/04_listmode_algorithms/00_run_listmode_mlem_osem_svrg.py
@@ -49,11 +49,6 @@ variants (sinogram MLEM, LM-MLEM, sinogram OSEM, LM-OSEM) because every
 objective implements the :class:`.C1Function` interface with the correct
 gradient w.r.t. :math:`x`.
 
-.. note::
-
-   To run this example locally, download
-   `example_utils.py <https://raw.githubusercontent.com/KUL-recon-lab/parallelproj/main/docs/examples/example_utils.py>`_
-   into the **same folder** as this script. Make sure ``parallelproj`` is installed.
 """
 
 # %%
@@ -61,8 +56,9 @@ from __future__ import annotations
 from collections.abc import Sequence
 from copy import copy
 from array_api_compat import size
-from example_utils import show_vol_cuts
-from example_utils import elliptic_cylinder_phantom
+
+from parallelproj._examples_utils import show_vol_cuts
+from parallelproj._examples_utils import elliptic_cylinder_phantom
 import numpy as np
 import matplotlib.pyplot as plt
 
@@ -80,7 +76,7 @@ from parallelproj.functions import (
 )
 
 # %%
-from example_utils import suggest_array_backend_and_device
+from parallelproj._examples_utils import suggest_array_backend_and_device
 
 # To use a specific backend and/or device, replace the None arguments, e.g.:
 #   xp, dev = suggest_array_backend_and_device(backend="numpy", dev="cpu") or by setting xp and dev manually